@@ -2,7 +2,7 @@
   "name": "next-wade-hammes",
   "version": "0.0.1",
   "private": true,
-  "packageManager": "pnpm@10.13.1",
+  "packageManager": "pnpm@10.14.0",
   "scripts": {
     "dev": "NODE_OPTIONS='--inspect' next dev -p 4431",
     "build": "next build && make sitemap",
@@ -13,47 +13,39 @@
     "lint": "biome check",
     "lint:ci": "biome ci --reporter=github",
     "lint:fix": "biome check --fix",
-    "lint:css": "stylelint ./src/**/*.css",
     "test:ci": "jest --passWithNoTests",
     "types:contentful": "export $(cat .env.local | sed '/^#/d; s/[$\"]//g'); cf-content-types-generator -s $CONTENTFUL_SPACE_ID -t $CONTENTFUL_CMA_TOKEN -o src/contentful/types -X && biome format --write src/contentful/types"
   },
   "dependencies": {
-    "@emotion/is-prop-valid": "^1.3.1",
-    "@jest/types": "^30.0.5",
-    "@next/third-parties": "^15.4.3",
+    "@next/third-parties": "^15.5.2",
     "@svgr/webpack": "^8.1.0",
-    "@testing-library/react": "^16.3.0",
+    "@types/webpack": "^5.28.5",
     "classnames": "^2.5.1",
+    "critters": "^0.0.25",
     "culori": "^4.0.2",
     "gsap": "^3.13.0",
-    "html-react-parser": "^5.2.6",
-    "next": "^15.4.3",
+    "next": "^15.5.2",
     "polished": "^4.3.1",
-    "postcss-styled-syntax": "^0.7.1",
-    "react": "^19.1.0",
-    "react-dom": "^19.1.0",
+    "react": "^19.1.1",
+    "react-dom": "^19.1.1",
     "react-intersection-observer": "^9.16.0",
     "tslib": "^2.8.1"
   },
   "devDependencies": {
-    "@biomejs/biome": "^2.1.2",
-    "@next/bundle-analyzer": "^15.4.3",
-    "@testing-library/jest-dom": "^6.6.3",
+    "@biomejs/biome": "^2.2.3",
+    "@jest/types": "^30.0.5",
+    "@next/bundle-analyzer": "^15.5.2",
+    "@testing-library/jest-dom": "^6.8.0",
+    "@testing-library/react": "^16.3.0",
     "@types/jest": "^30.0.0",
-    "@types/node": "^24.1.0",
-    "@types/react": "^19.1.8",
-    "@types/webpack": "^5.28.5",
-    "critters": "^0.0.25",
+    "@types/node": "^24.3.1",
+    "@types/react": "^19.1.12",
     "csstype": "^3.1.3",
-    "jest": "^30.0.5",
-    "jest-environment-jsdom": "^30.0.5",
-    "react-is": "^19.1.0",
-    "react-test-renderer": "^19.1.0",
-    "stylelint": "^16.22.0",
-    "stylelint-config-recommended": "^16.0.0",
-    "stylelint-webpack-plugin": "^5.0.1",
-    "ts-jest": "^29.4.0",
+    "jest": "^30.1.3",
+    "jest-environment-jsdom": "^30.1.2",
+    "stylelint-config-recommended": "^17.0.0",
+    "ts-jest": "^29.4.1",
     "ts-node": "^10.9.2",
-    "typescript": "^5.8.3"
+    "typescript": "^5.9.2"
   }
 }
